% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_Realizations.R
\name{Add_Realizations}
\alias{Add_Realizations}
\title{Adds the results of each bets.}
\usage{
Add_Realizations(path)
}
\arguments{
\item{path}{the path used by the webscrapper to save data}
}
\description{
This function loops over the avaible data and adds the outcomes of each avaible bet.
}
