% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arbitrage.R
\name{arbitrage}
\alias{arbitrage}
\title{Initiate or update the arbitrage database}
\usage{
arbitrage(path)
}
\arguments{
\item{path}{the path used by the webscrapper to save data}
}
\description{
This function creates or updates the past and present arbitrage opportunities
}
