% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_temp_file_Covariances.R
\name{compute_covariance_matrices}
\alias{compute_covariance_matrices}
\title{Compute covariance factors}
\usage{
compute_covariance_matrices(path)
}
\arguments{
\item{path}{the path used by the webscrapper to save data}
}
\description{
This function computes E\link{I_1 * I_2} / (E\link{I_1} E\link{I_2}), which is to be used to estimate the covariances when computing portfolio weights
}
