% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute_Fair_Factors.R
\name{Compute_Fair_Factors}
\alias{Compute_Fair_Factors}
\title{Performs simulations and returns the estimated probabilies for avaible bets.}
\usage{
Compute_Fair_Factors(path, n_simulations = NULL, ncores = NULL)
}
\arguments{
\item{path}{the path used by the webscrapper to save data}

\item{n_simulations}{numbers of simulations}

\item{ncores}{numbers of cores to be used during the simulations}
}
\description{
This function computes the estimates needed to identify arbitrage opportunities.
}
