% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_Markov_Chain_Data.R
\name{Extract_Markov_Chain_Data}
\alias{Extract_Markov_Chain_Data}
\title{Extract and process data needed to set up Markov Chains}
\usage{
Extract_Markov_Chain_Data(date, lineup, params, id = NULL)
}
\arguments{
\item{date}{the date at which the match takes (or took) place}

\item{lineup}{a list of size 3 containing: home/away = list(bat = ..., pitch = ...), teams = list(home = ..., away =...). The most inner elements are vectors containing characters, i.e.: names.}

\item{params}{a list of size 3 containing: n_season = ..., n_match_bat = ..., n_match_pitch = ... . These are the time-windows used to compute averages.}

\item{id}{the match ID. Defaults to NULL; if specified, the lineup will be retrieved from historical files. (I.e.: use IDs to build a design matrix X post-webscrapping.)}
}
\value{
A single row of the design matrix X given two teams and two rosters.
}
\description{
This function returns transition and event matrices, condition points distribution,
batting orders, ect.
}
\details{
Note: you need to initiate DB (.../MLB_Modeling/Scores/Clean_Data/Seasonal_Database.rds),
scores (/MLB_Modeling/Scores/Clean_Data/DB_Scores.rds) and play_by_play
("/MLB_Modeling/Scores/Clean_Data/Markov_Database.rds") prior to running this function.
}
