% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_MiseOjeu_data.R
\name{process_MiseOjeu_data}
\alias{process_MiseOjeu_data}
\title{Initiate or update the bets database}
\usage{
process_MiseOjeu_data(path)
}
\arguments{
\item{path}{the path used by the webscrapper to save data}
}
\description{
This function creates the database with the bets offered by Loto-Quebec
}
