% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_Pinnacle_data.R
\name{process_Pinnacle_data}
\alias{process_Pinnacle_data}
\title{Initiate or update the bets database}
\usage{
process_Pinnacle_data(path)
}
\arguments{
\item{path}{the path used by the webscrapper to save data}
}
\description{
This function creates the database with the bets offered by Pinnacle (half innings and game, select bets)
}
