% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_PlayByPlay_data.R
\name{process_PlayByPlay_data}
\alias{process_PlayByPlay_data}
\title{Initiate or update the play-by-play database}
\usage{
process_PlayByPlay_data(path)
}
\arguments{
\item{path}{the path used by the webscrapper to save data}
}
\description{
This function creates the database with the play-by-play data
}
