% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Setup_Markov_Database.R
\name{Setup_Markov_Database}
\alias{Setup_Markov_Database}
\title{Process play-by-play data}
\usage{
Setup_Markov_Database(path)
}
\arguments{
\item{path}{the path used by the webscrapper to save data}
}
\description{
This function creates the database needed to set up Markov chains.
}
