% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmulate_inning.R
\name{simmulate_game}
\alias{simmulate_game}
\title{Simulates an entire game}
\usage{
simmulate_game()
}
\value{
A simmulated game
}
\description{
This function simmulates an entire game
}
\details{
Note: you need to initiate "markov" with Extract_Markov_Chain_Data(...)
}
