% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmulate_inning.R
\name{simmulate_inning}
\alias{simmulate_inning}
\title{Simulates an inning}
\usage{
simmulate_inning(inn, team)
}
\arguments{
\item{inn}{The inning}

\item{team}{home or away}
}
\value{
A simmulated inning
}
\description{
This function simmulates a single inning
}
\details{
Note: you need to initiate "markov" with Extract_Markov_Chain_Data(...)
}
