% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_database.R
\name{update_database}
\alias{update_database}
\title{Updating the score and player data}
\usage{
update_database(path)
}
\arguments{
\item{path}{the path used by the webscrapper to store data}
}
\description{
This function updates the score and player statistics R databases.
}
