% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_design_matrix.R
\name{update_design_matrix}
\alias{update_design_matrix}
\title{Update the regression data}
\usage{
update_design_matrix(
  scrapper_path,
  avg_number,
  params = list(n_season = 2, n_match_bat = 120, n_match_pitch = 120)
)
}
\arguments{
\item{scrapper_path}{the path used by the webscrapper to save data}

\item{avg_number}{a tag used to save file (i.e.: to distinguish between different time windows used to compute averages)}
}
\description{
This function updates the design matrix, the score and the bookmakers data needed for regression purposes
}
