% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_design_matrix.R
\name{update_design_matrix}
\alias{update_design_matrix}
\title{Update the regression data}
\usage{
update_design_matrix(path, version)
}
\arguments{
\item{path}{the path used by the webscrapper to save data}

\item{version}{a tag used to save file (i.e.: to distinguish between different time windows used to compute averages)}
}
\description{
This function updates the design matrix, the score and the bookmakers data needed for regression purposes
}
